refactor: rename type.ts
