+++ src/renderer/src/components/connection-wizard/wizard-modal.tsx
@@ -2,7 +2,7 @@ import { JSX, useEffect, useState } from 'react'
 import { toast } from 'sonner'
 import { FooterButtons } from './footer-buttons'
 import { SelectDatabase } from './select-database'
-import { ConnectionDeatil, DatabaseInfo, DBSetupStep } from './type'
+import { ConnectionDeatil, DatabaseInfo, DBSetupStep } from './wizard.type'
 import Sidebar from './sidebar'
 import InstallDriver from './install-driver'
 import EnterConnectionDetails from './enter-connection-details'
