+++ src/renderer/src/components/connection-wizard/wizard.type.ts
